pyproject.toml
[project]
name = "all-in-one-ai-assistant"
version = "1.0.0"
description = "An All-in-One AI Assistant using Google's Gemini API"
authors = [{ name = "Padam Srishank", email = "your.email@example.com" }]
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "streamlit",
    "google-generativeai",
    "requests",
]
